move xgboost.validate into top level
